APSR-1418: minor tidy up
