+++ app/uk/gov/hmrc/apiplatformmicroservice/apidefinition/services/FilterApis.scala
@@ -108,13 +108,21 @@ trait FilterGateKeeperSubscriptions extends FilterApis {
   }
 }
 
-trait FilterForCompinedApis extends FilterApis {
+trait FiltersForCompinedApis extends FilterApis {
   private def isOnlyPublicAccess(v: ExtendedApiVersion): Boolean ={
     (v.productionAvailability, v.sandboxAvailability) match {
       case (Some(prod: ApiAvailability), Some(sand: ApiAvailability)) =>
         isPublicAccess(prod.access) && isPublicAccess(sand.access)
       case _ => false
     }
+  }
+
+  def filterOutRetiredApis(definitions: List[ApiDefinition]): List[ApiDefinition] = {
+    def filterOutRetiredVersions(definition: ApiDefinition): Option[ApiDefinition] = {
+      val filteredVersions = definition.versions.filterNot(_.status == ApiStatus.RETIRED)
+      if(filteredVersions.isEmpty) None else Some(definition.copy(versions = filteredVersions))
+    }
+    definitions.flatMap(filterOutRetiredVersions)
   }
 
   def allVersionsArePublicAccess(a: ApiDefinition): Boolean = a.versions.forall(v => isPublicAccess(a.context, v))
